+++ internal/config/config.go
@@ -113,6 +113,9 @@ type PackageList struct {
 type DirConfig struct {
 	Profile  string            `yaml:"profile" toml:"profile"`
 	Targets  map[string]string `yaml:"targets" toml:"targets"`
+	Owner    string            `yaml:"owner" toml:"owner"`
+	Group    string            `yaml:"group" toml:"group"`
+	Perm     string            `yaml:"perm" toml:"perm"`
 	Packages *PackageList      `yaml:"packages" toml:"packages"`
 	Scripts  *DirScripts       `yaml:"scripts" toml:"scripts"`
 }